Fix bug #6 and mismatch between lon/lat and x/y coord names
The GMT function grdcut doesn't seem to like integer valued arrays. This
has been fixed by converting all input netcdf files to float array
during loading. Additionally, the coordinate system seems to have been
renamed from lon/lat to x/y at some point. This has been fixed.

+++ loaders.py
@@ -20,7 +20,7 @@ def load_source(filepath, plot=False):
         raise RuntimeError(f"Error: filetype {ext} not recognised.")
 
 def load_netcfd(filepath, plot=False, convert_to_xyz=False):
-    xr_data = xr.open_dataarray(filepath)
+    xr_data = xr.open_dataarray(filepath).astype('float32')
 
     print(f"Resolution: {xr_data.values.shape}")
